Changes validator interface to abstract class.
